refactor: 컴포넌트 재분리 하면서 파일 구조 변경
