+++ src/router/AuthRoutes.tsx
@@ -10,7 +10,7 @@ import TypeSelectPage from "@/pages/auth/TypeSelectPage";
 import BusinessVerificationPage from "@/pages/auth/BusinessVerification";
 import OwnerTermsPage from "@/pages/auth/OwnerTermsPage";
 import RecruitWriteContainer from "@/pages/owner/Recruit";
-import Review from "@/pages/owner/Review";
+import ReviewManagePage from "@/pages/owner/ReviewManagePage";
 const AuthRoutes: RouteObject[] = [
     {
         path: "/",
@@ -60,7 +60,7 @@ const AuthRoutes: RouteObject[] = [
     },
     {
         path: "owner/userinfo/reviews",
-        element: <Review />,
+        element: <ReviewManagePage />,
     },
 ];
